+++ ChatClient/src/main/java/module-info.java
@@ -3,9 +3,12 @@ module com.chatapp.client {
     requires javafx.fxml;
     requires javafx.media;
     requires javafx.web;
+    requires javafx.swing;
     requires com.google.gson;
     requires org.slf4j;
     requires emoji.java;
+    requires webcam.capture;
+    requires java.desktop;
 
     opens com.chatapp.client to javafx.fxml;
     opens com.chatapp.client.controller to javafx.fxml;